build(bundle): add `sourceMappingURL` in minified code
This is because otherwise sourcemaps will not be mapped for UMD bundles
