+++ tools/make-closure-core.js
@@ -1,15 +1,3 @@
-var compiler = require('google-closure-compiler-js').compile;
-var fs = require('fs');
+var makeClosure = require('./make-closure');
 
-var source = fs.readFileSync('dist/global/rxjs.umd.js', 'utf8');
-
-var compilerFlags = {
-  jsCode: [{src: source}],
-  languageIn: 'ES2015',
-  createSourceMap: true,
-};
-
-var output = compiler(compilerFlags);
-
-fs.writeFileSync('dist/global/rxjs.umd.min.js', output.compiledCode, 'utf8');
-fs.writeFileSync('dist/global/rxjs.umd.min.js.map', output.sourceMap, 'utf8');
+makeClosure('dist/global/rxjs.umd.js');
